remove most "blitz db" commands, leave just "blitz seed"
